requirements.txt
fastapi>=0.68.0
uvicorn>=0.18.3
pydantic>=1.10.0
python-dotenv>=0.19.0
httpx>=0.23.0
pydantic_settings>=0.2.0
sqlalchemy>=2.0.0
alembic>=1.7.0
psycopg2-binary >=2.9.0
geopy >=2.0.0
google-auth-oauthlib>=0.4.6
google-api-python-client>=2.0.0
pytz >=2024.1
boto3 >=1.35.0
python-multipart >=0.0.9